itest: give shared fixtures neutral ownership
The funding amounts, confirmation bound, lease duration and unknown
outpoint were owned by the UtxoManager test file, so any other
component's tests had to depend on that file to reuse them. Move the
shared values to a file no component owns.
